Move all unnecessary files to new folders
