+++ examples/eeg_video_sync/index.py
